Reduce code duplication in computational designs.
